+++ src/test/java/unit/bswap/TestClaimRewards.java
@@ -1,0 +1,41 @@
+package unit.bswap;
+
+import com.binance.connector.client.enums.HttpMethod;
+import com.binance.connector.client.impl.SpotClientImpl;
+import okhttp3.mockwebserver.Dispatcher;
+import okhttp3.mockwebserver.MockWebServer;
+import org.junit.Before;
+import org.junit.Test;
+import unit.MockWebServerDispatcher;
+
+import java.util.LinkedHashMap;
+
+import static org.junit.Assert.assertEquals;
+
+public class TestClaimRewards {
+    private MockWebServer mockWebServer;
+    private String baseUrl;
+    private final String prefix = "/";
+    private final String MOCK_RESPONSE = "{\"key_1\": \"value_1\", \"key_2\": \"value_2\"}";
+    private final String apiKey = "apiKey";
+    private final String secretKey = "secretKey";
+
+    @Before
+    public void init() {
+        this.mockWebServer = new MockWebServer();
+        this.baseUrl = mockWebServer.url(prefix).toString();
+    }
+
+    @Test
+    public void testClaimRewards() {
+        String path = "/sapi/v1/bswap/claimRewards";
+        LinkedHashMap<String,Object> parameters = new LinkedHashMap<>();
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(prefix, path, MOCK_RESPONSE, HttpMethod.POST, 200);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(apiKey, secretKey, baseUrl);
+        String result = client.createBswap().claimRewards(parameters);
+        assertEquals(MOCK_RESPONSE, result);
+    }
+}